requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
Flask==1.1.2
get==2019.4.13
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.5
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
typing-extensions==3.7.4.2
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2